requirements.txt
numpy==1.24.4
scipy
pandas
matplotlib
open3d==0.12.0
catkin_pkg
scikit-learn==1.3.2
pybullet==2.7.9
pytorch-ignite
tensorboard
tqdm
trimesh
urchin
cython
scikit-image
pykdtree
ninja
theseus-ai
opencv-python
einops
Cython